fix: run prettier (#143)
* fix: prettier configuration

* fix: format

+++ packages/greenproof-contracts/contracts/facets/VotingFacet.sol
@@ -170,7 +170,7 @@ contract VotingFacet is IVoting, IReward {
         }
 
         votes = new bytes32[](numberOfVotes);
-        for (uint i; i < numberOfVotes; i++) {
+        for (uint256 i; i < numberOfVotes; i++) {
             votes[i] = votesContainer[i];
         }
     }
